# perf(es/minifier): Improve codegen of name mangler